requirements.txt
telethon>=1.34.0
ccxt>=4.3.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn>=0.27.0
psycopg2-binary>=2.9.9
